requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.35
bcrypt==4.2.0
cryptography==41.0.7
pyotp==2.9.0
pydantic
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
jinja2==3.1.2
qrcode==7.4.2
pillow
passlib==1.7.4
argon2-cffi==23.1.0
pydantic-settings

